Change Generator return type from int to float

+++ 0x02-python_async_comprehension/0-async_generator.py
@@ -6,7 +6,7 @@ from typing import Generator
 import random
 
 
-async def async_generator() -> Generator[int, None, None]:
+async def async_generator() -> Generator[float, None, None]:
     """wait 1 second, then yield a random number between 0 and 10"""
 
     for i in range(10):
